reorganize events page components
